Feat: #23 - place list 컴포넌트 분리
