+++ pages/search/[result].tsx
@@ -1,5 +1,6 @@
 import Button from '@/components/common/Button';
 import SearchBar from '@/components/common/SearchBar';
+import PlaceItem from '@/components/placeList/PlaceItem';
 import COLOR from '@/constants/colors';
 import FONT from '@/constants/fonts';
 import { GetServerSideProps } from 'next';
@@ -44,32 +45,8 @@ const SearchResultIndex = ({ result }: { result: string }) => {
     <>
       <SearchBar text={result} />
       <div>
-        {searchList.map((result) => (
-          <SearchListWrapper key={result.id}>
-            <LeftWrapper>
-              <PlaceHeadWrapper>
-                <PlaceName style={FONT.HEADLINE2}>{result.name}</PlaceName>
-                <PlaceType style={FONT.BODY2} type={result.type}>
-                  {result.type}
-                </PlaceType>
-              </PlaceHeadWrapper>
-              <PlaceLocation style={FONT.BODY2}>{result.location}</PlaceLocation>
-              <PlaceTimeWrapper style={FONT.BODY2}>
-                <div style={{ color: COLOR.RED }}>운영종료</div>
-                <div>{result.startTimeAt} 에 운영시작</div>
-              </PlaceTimeWrapper>
-              <IconWrapper>
-                <Image src='/images/wheelChair.svg' width={30} height={30} />
-                <Image src='/images/elevator.svg' width={30} height={30} />
-                <Image src='/images/slope.svg' width={30} height={30} />
-              </IconWrapper>
-            </LeftWrapper>
-            <RightWrapper>
-              <Button bgColor={COLOR.BLUE1} color={COLOR.WHITE}>
-                길찾기
-              </Button>
-            </RightWrapper>
-          </SearchListWrapper>
+        {searchList.map((place) => (
+          <PlaceItem place={place} />
         ))}
       </div>
     </>
@@ -83,65 +60,5 @@ export const getServerSideProps = async ({ query: { result } }: { query: { resul
     props: { result }
   };
 };
-
-const IconWrapper = styled.div`
-  display: flex;
-  flexbasis: 20%;
-  gap: 5px;
-`;
-
-const LeftWrapper = styled.div`
-  display: flex;
-  flex-basis: 80%;
-  flex-direction: column;
-  gap: 10px;
-  padding: 20px;
-`;
-
-interface PlaceTypeProps {
-  type: string;
-}
-
-type ObjType = {
-  [index: string]: string;
-};
-
-const TYPE_TO_COLOR: ObjType = {
-  박물관: COLOR.ORANGE,
-  미술관: COLOR.GREEN,
-  전시회: COLOR.RED
-};
-
-const PlaceType = styled.div<PlaceTypeProps>`
-  color: ${(props) => TYPE_TO_COLOR[props.type]};
-`;
-
-const PlaceHeadWrapper = styled.div`
-  display: flex;
-  gap: 10px;
-`;
-
-const PlaceName = styled.div``;
-
-const PlaceLocation = styled.div`
-  color: ${COLOR.GREY};
-`;
-
-const PlaceTimeWrapper = styled.div`
-  display: flex;
-  gap: 10px;
-  color: ${COLOR.GREY};
-`;
-
-const RightWrapper = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-const SearchListWrapper = styled.div`
-  display: flex;
-  border-bottom: 1px solid black;
-  border-color: ${COLOR.LINE};
-`;
 
 export default SearchResultIndex;
